round to intger for legend

+++ rp_plot.py
@@ -50,7 +50,7 @@ def plotrh(hist,showbpm=True):
         mod_freq_res = 1.0 / (2**18/44100.0)
         plot_index = range(0, hist.shape[0] + 1, 5)  # 5 = step for xticks
         plot_base = np.array(plot_index) + 1  # + 1 otherwise we start we 0 bpm
-        bpm = np.around(plot_base * mod_freq_res * 60, 0)
+        bpm = np.around(plot_base * mod_freq_res * 60, 0).astype(int)  # integer for legend
         plt.xticks(plot_index, bpm)
         plt.xlabel('bpm')
     else:
